Simplify theta glm
bo is redundant and allowing yearly variation in theta (b2.theta) masks yearly variation in abundance. Probably should do theta as a heir beta but the point estimates are similar and I don’t want to change the function that depend on theta for AGAIN.  wait till other parts of the model are more settled.