clean 2022/create 2023 estimate
Mostly added new data. Significant notes are 1) the model converges poorly but mostly for age3 fish, 2) the Yentna stock comp keeps moving w the addition of new sonar data. Not sure if this is a feature or a bug. Matt sent an email on 10272023 with a posterior check against this issue.


Former-commit-id: 1d5d7b029fdd24895bb943944203f28fa40f15e7